Revert "Milestone 2: Add, Edit and Remove functionalities"
