+++ src/index.html
@@ -7,10 +7,10 @@
   </head>
   <body>
     <main><div class="listContainer">
-      <div class="title"><h3>Today's To Do</h3><img class = 'imgRefresh' alt = 'refresh' src = '../src/refresh.png'></div>
+      <div class="title"><h3>Today's To Do</h3><img class = 'imgRefresh' alt = 'refresh'></div>
       <form class = "addTask">
-        <input type="text" placeholder="Add to your list..." class = 'add'>
-        <input type="submit" value = '' class = "submit"><img class = 'imgSubmit' alt = 'return'  src = '../src/return.png'>
+        <input type="text" placeholder="Add to your list...">
+        <input type="submit" value = '' class = "submit"><img class = 'imgSubmit' alt = 'return'>
       </form>
       <ul class="to-do-list"></ul>
       <div class="clear">Clear all completed</div>
