+++ src/index.js
@@ -1,43 +1,35 @@
 import './style.css';
+import moveSrc from './move.png';
 import refreshSrc from './refresh.png';
 import returnSrc from './return.png';
-import {
-  addTask,
-  list,
-} from './add.js';
-import clearTask from './remove.js';
-import appendTask from './addTaskUI.js';
-import Task from './tasks.js';
 
-const storage = JSON.parse(localStorage.getItem('tasks')) || [];
-const addUI = document.querySelector('.add');
-const deleteUI = document.querySelector('.clear');
 const returnImg = document.querySelector('.imgSubmit');
-const refreshImg = document.querySelector('.imgRefresh');
-
-storage.forEach((item) => {
-  const task = new Task(item.description, item.completed, item.index);
-  list.push(task);
-  appendTask(task, list);
-});
-
 returnImg.src = returnSrc;
+
+const refreshImg = document.querySelector('.imgRefresh');
 refreshImg.src = refreshSrc;
 
-addUI.addEventListener('change', () => {
-  addTask(addUI.value, false, list.length + 1);
-});
+const toDoListUI = document.querySelector('.to-do-list');
 
-refreshImg.addEventListener('click', () => {
-  document.location.href = './index.html';
-});
-
-deleteUI.addEventListener('click', () => {
-  const checks = document.querySelectorAll('.check');
-  for (let i = 0; i < checks.length; i += 1) {
-    if (checks[i].checked) {
-      clearTask(checks[i].value, list);
-    }
+class Task {
+  constructor(description, completed, index) {
+    this.description = description;
+    this.completed = completed;
+    this.index = index;
   }
-  localStorage.setItem('tasks', JSON.stringify(list));
-});
+}
+
+const list = [new Task('Finsh the project', false, 0), new Task('Wash the dishes', false, 1), new Task('Go on zoom meeting', false, 2)];
+
+const appendTask = (item) => {
+  const taskUI = document.createElement('li');
+  taskUI.innerHTML = `<label><input type='checkbox' id = task-${item.index} class = 'check' value=${item.index}>${item.description}</label>`;
+  const moveImg = new Image();
+  moveImg.src = moveSrc;
+  taskUI.appendChild(moveImg);
+  toDoListUI.appendChild(taskUI);
+};
+
+for (let i = 0; i < list.length; i += 1) {
+  appendTask(list[i]);
+}
