--- src/remove.js
@@ -1,20 +1,0 @@
-const clearTask = (index, list) => {
-  if (list.length === 0) return;
-
-  list.splice(index - 1, 1);
-
-  document.querySelector(`#task-${index}`).parentNode.parentNode.remove();
-  localStorage.setItem('tasks', JSON.stringify(list));
-
-  if (list.length === 0) return;
-
-  for (let i = index - 1; i < list.length; i += 1) {
-    const nextCheck = document.querySelector(`#task-${i + 2}`);
-    nextCheck.id = `task-${i + 1}`;
-    nextCheck.value -= 1;
-    list[i].index -= 1;
-    localStorage.setItem('tasks', JSON.stringify(list));
-  }
-};
-
-export default clearTask;
