--- src/tasks.js
@@ -1,8 +1,0 @@
-class Task {
-  constructor(description, completed, index) {
-    this.description = description;
-    this.completed = completed;
-    this.index = index;
-  }
-}
-export default Task;
